+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to hp (hittyping) will be documented in this file.
 
-## [Unreleased]
+## [0.7.7] - 2026-01-25
 
 ### Added